Stop Jenkins service before changing config data
After installation, Jenkins must be started up so that it can create
the files and directories that it needs to run. However, Jenkins
should be stopped before changing any configuration data, since some
of the files generated by Jenkins (for example, hudson.util.Secret)
are also mapped in memory and flushed to disk when Jenknis shuts down,
which overwrites any custom user data in the process.