Safer handling of secrets data during configuration (#8)
* Create intermediate dirs for custom files

If the user specifies a custom file which lives in a subdirectory of
JENKINS_HOME then unfortunately the "Configure custom files" task will
fail because the template module will not create the parent directory
for the destination file.

* Move plugin configuration tasks to new file

Jenkins should be stopped when touching configuration data, but it
needs to be running to install plugins. This change splits up the base
and plugin configuration into two separate files so that the service
can be in the correct state for each one.

* Stop Jenkins service before changing config data

After installation, Jenkins must be started up so that it can create
the files and directories that it needs to run. However, Jenkins
should be stopped before changing any configuration data, since some
of the files generated by Jenkins (for example, hudson.util.Secret)
are also mapped in memory and flushed to disk when Jenknis shuts down,
which overwrites any custom user data in the process.

* Set correct permissions on the Jenkins secrets dir

* Create Jenkins user/group explicitly

The "Ensure jenkins home dir is created" task can fail if the Jenkins
user/group don't yet exist.