Use uri task to wait for Jenkins to start up
Using the pause task is a bit problematic, since it can either waste a
lot of time on fast systems or time out on really slow ones. This
commit uses the uri task to fetch the Jenkins homepage until it
returns status code 200. Jenkins will return 503 (service unavailable)
while it is loading.