Put Jenkins in quiet mode when installing plugins/jobs
Jenkins needs to be up and running in order to install plugins, but
it's a bad idea to accept new builds, since any number of things can
go wrong during this time.

Accepting builds while copying job configuration files is equally
dangerous. This could be avoided by shutting down Jenkins, as we do
when configuring Jenkins core files.  However, as long as we're
already in quiet mode for the plugin installation we can just keep
Jenkins running and avoid a potentially long service restart.