Don't restart Jenkins at end of apt install task
Any tasks run after this one should be responsible for putting the
Jenkins service in the state they need. Also, the next task stops
Jenkins immediately (though this task shouldn't need to care about
that), so it's a bit wasteful to restart the service just to see it
stopped again a few seconds later.