Improve service lifecycle (#39)
* Add task for starting Jenkins

The introduction of this task will assist in eliminating both the
restart.yml and wait-for-start.yml tasks. This task handles both
starting the Jenkins service and waiting for it to complete starting
(formerly handled by wait-for-start.yml). In this manner, we can
achieve a more fine-grained control of the Jenkins service status.

* Put Jenkins in correct state in task files

Previous to this change, main.yml attempted to manage Jenkins' service
state before including any of the other task files. This got a bit
messy and also led to some cases of unnecessary service restarts.

Now, each task file will place Jenkins in the desired state before
attempting to do any work. This guarantees that no matter what state
the service is in when a task finishes, the next task will execute
without problems. This also makes it possible for playbooks including
this role to be able to use `include_tasks`/`tasks_from` from this
role without needing to worry about putting the Jenkins service in the
correct state beforehand.

* Don't restart Jenkins at end of apt install task

Any tasks run after this one should be responsible for putting the
Jenkins service in the state they need. Also, the next task stops
Jenkins immediately (though this task shouldn't need to care about
that), so it's a bit wasteful to restart the service just to see it
stopped again a few seconds later.

* Move custom file tasks to own file

Custom files should be handled after plugin installation. In case the
user has any custom files related to plugins, those plugins should be
present on Jenkins first, otherwise there is a risk that they will
promptly overwrite these files with default values.

* Take Jenkins out of quiet mode after tasks finish

In case the user does something potentially disruptive in a custom
jenkins_after_config_jobs_file task, we should make sure that Jenkins
isn't accepting new builds just yet.

* Use wait_for instead of pause

The pause module blocks parallel execution for other hosts, since it
is intended to gather user input. We are only interested in sleeping a
few seconds to wait for the service to be in the intended state, so
the wait_for is a better fit.

* Include jenkins_after_config_jobs_file if defined

This commit defines a default value for
jenkins_after_config_jobs_file, as well as providing a brief
description of the variable. This makes `empty.yml` obsolete.

* Start Jenkins after all tasks are complete

Of course, this must also be done before cancelling quiet mode.

* Remove unused task files

* Don't reload Jenkins after configuring jobs

Jenkins will be stopped before touching the jobs, and started up
afterwards either by `main.yml` or another task which requires it to
be started. Therefore it is not necessary to call reload, since
Jenkins will pick up on the new jobs when it restarts.

* Cleanup: add newlines between tasks in main.yml