Ensure correct ownership of JENKINS_HOME directory (#38)
For the native installation types (apt, yum), JENKINS_HOME is created
with the Jenkins user on when the package is installed. In either
case, it is quite likely that root will own this directory. Therefore,
we should ensure that the correct Jenkins user owns this directory.

In the case of the user module, the home directory created by Ansible
is always owned by root:root, which is a known bug in Ansible:

https://github.com/ansible/ansible/issues/24862

As luck would have it, the "Create intermediate dirs for custom files"
task was accidentally doing this work for us. However, if the user did
not set any items for the jenkins_custom_files var, then this would
cause Jenkins to fail at startup.