Introduce variable for plugin installation timeout (#42)
Some plugins, such as Blue Ocean, have large dependency trees and can
take some time to install. This commit introduces a new variable,
which can be increased for users who are installing a large amount of
plugins, plugins with deep dependency trees, or perhaps are using a
Jenkins master with meager I/O capabilities.

The default value is 300, which is the same as the default value for
the timeout option of the wait_for module.