Add Molecule tests to repository
This commit introduces Molecule, a framework for testing Ansible
roles, into the repository. The testing is somewhat limited at the
moment; it only tests the apt installation method with Ubuntu 16.04
Xenial, and only tests a few role features, namely:

- Whether Jenkins was successfully installed
- Whether the Jenkins service started up
- Installation of a sample job
- Installation of a plugin

It would be nice to expand test and platform coverage of this in the
future, but this commit at least gives us a base from which we can
expand.