Add Molecule tests, Travis CI build (#46)
* Add jenkins_home option to jenkins_plugin

This fixes an error with "Could not replace file" if the Jenkins user
doesn't have permission to write to /var/lib/jenkins (the default
value of this option).

* Add .yamllint configuration file

* Fix yamllint warning for spaces before comments

* Remove extra newlines

* Remove extraneous comments from meta YAML file

* Fix indentation in meta YAML file

* Use true/false for truthy values

This fixes an ansible-lint warning

* Add names to all tasks

This fixes an ansible-lint violation

* Fix indentation

* Make tasks idempotent

This fixes an ansible-lint warning as well as a test failure in
molecule. In some cases, the warning has just been suppressed with a
noqa comment -- in the future, we should fix these tasks and make them
actually idempotent (if possible).

* Break long lines

* Improve YAML readability

* Add Molecule tests to repository

This commit introduces Molecule, a framework for testing Ansible
roles, into the repository. The testing is somewhat limited at the
moment; it only tests the apt installation method with Ubuntu 16.04
Xenial, and only tests a few role features, namely:

- Whether Jenkins was successfully installed
- Whether the Jenkins service started up
- Installation of a sample job
- Installation of a plugin

It would be nice to expand test and platform coverage of this in the
future, but this commit at least gives us a base from which we can
expand.

* Add .travis.yml to repository