Improve authentication mechanisms (#50)
* Configure files before plugins

This is necessary because certain files, such as user config.xml
files, may contain API tokens which are needed for authenticated
plugin installation. Therefore these files must be in place before
attempting to start up Jenkins for the first time.

* Install custom plugins first

Since the custom plugins need to be installed by Jenkins when it
starts up, this minor optimization saves a bit of time during plugin
installation.

* Add jenkins_auth variable to determine auth type

This change is the first step in flexibly supporting multiple
authentication types to use when deploying Jenkins. In a future
commit, support for API tokens will be added.

This change also renames jenkins_token to jenkins_crumb_token, in
order to be more explicit about the token type.

* Save the web session cookie along with the crumb

As of Jenkins 2.176.2, Jenkins' CSRF issuer requires a corresponding
web session ID to be submitted along with the crumb. For more info,
see:

https://jenkins.io/security/advisory/2019-07-17/#SECURITY-626

* Support API token authentication

* Add a molecule scenario test for API tokens

This is hopefully the second of many additional non-sanity tests that
will be added to ensure that different behaviors of this role are
working properly.

* Add some tasks to do basic sanity checking of vars

This could probably be expanded in the future for other states that we
need to verify before beginning a deployment.

* Add documentation on authentication and security