Add support for HTTPS, improve handling of Jenkins URL (#48)
* Move duplicated tasks to common file

* Set the jenkins_url fact based on protocol

This change un-defines jenkins_url in the default's main.yml file, and
instead allows users to define jenkins_hostname instead. The rationale
for this is:

- Previously jenkins_url was not a true URL as it did not include the
  port. This was added by the role as necessary.
- It required the user to specify the protocol, which became a bit
  awkward when adding support for HTTPS.

This change includes backwards-compatibility support for playbooks
which define jenkins_url by re-defining it to the new value with a
port, and displaying a debug warning message to the user.

* Add support for HTTPS

* Improve "Role Variables" section of readme

This section originally copy/pasted the defaults/main.yml file, which
is not very sustainable since changes in that file were not always
propagated to the README. Instead, this section now includes
documentation on the most relevant options and links to the actual
file for everything else.